+++ buffer/buffer_get_token.c
@@ -2,8 +2,8 @@
 #include "buffer.h"
 #include "scan.h"
 
-int buffer_get_token(buffer* b,char* x,unsigned long int len,const char* charset,unsigned long int setlen) {
-  unsigned long int blen;
+ssize_t buffer_get_token(buffer* b,char* x,size_t len,const char* charset,size_t setlen) {
+  size_t blen;
 
   for (blen=0;blen<len;++blen) {
     register int r;
