+++ buffer/buffer_getn.c
@@ -1,11 +1,11 @@
 #include "byte.h"
 #include "buffer.h"
 
-int buffer_getn(buffer* b,char* x,unsigned long int len) {
-  unsigned int blen;
+ssize_t buffer_getn(buffer* b,char* x,size_t len) {
+  size_t blen;
 
   for(blen=0;blen<len;++blen) {
-    register int r;
+    register ssize_t r;
     if ((r=buffer_getc(b,x))<0) return r;
     if (r==0) break;
     ++x;
