+++ buffer/buffer_putflush.c
@@ -1,6 +1,6 @@
 #include "buffer.h"
 
-int buffer_putflush(buffer* b,const char* x,unsigned long int len) {
+int buffer_putflush(buffer* b,const char* x,size_t len) {
   if (buffer_put(b,x,len)<0) return -1;
   if (buffer_flush(b)<0) return -1;
   return 0;
