+++ case.h
@@ -1,15 +1,17 @@
 #ifndef CASE_H
 #define CASE_H
 
+#include <stddef.h>
+
 /* turn upper case letters to lower case letters, ASCIIZ */
 void case_lowers(char *s);
 /* turn upper case letters to lower case letters, binary */
-void case_lowerb(void *buf,unsigned long len);
+void case_lowerb(void *buf,size_t len);
 
 /* like str_diff, ignoring case */
 int case_diffs(const char *,const char *);
 /* like byte_diff, ignoring case */
-int case_diffb(const void *,unsigned long,const void *);
+int case_diffb(const void *,size_t ,const void *);
 
 /* like str_start, ignoring case */
 int case_starts(const char *,const char *);
