+++ fmt/fmt_long.c
@@ -1,6 +1,6 @@
 #include "fmt.h"
 
-unsigned int fmt_long(char *dest,long int i) {
+size_t fmt_long(char *dest,long int i) {
   if (i<0) {
     if (dest) *dest++='-';
     return fmt_ulong(dest,-i)+1;
