+++ fmt/fmt_minus.c
@@ -1,6 +1,6 @@
 #include "fmt.h"
 
-unsigned int fmt_minus(char *dest,int i) {
+size_t fmt_minus(char *dest,int i) {
   if (i<0) {
     if (dest) *dest='-';
     return 1;
