+++ fmt/fmt_plusminus.c
@@ -1,6 +1,6 @@
 #include "fmt.h"
 
-unsigned int fmt_plusminus(char *dest,int i) {
+size_t fmt_plusminus(char *dest,int i) {
   if (i) {
     if (dest) *dest=(i>=0?'+':'-');
     return 1;
