+++ fmt/fmt_str.c
@@ -1,6 +1,6 @@
 #include "fmt.h"
 
-unsigned long fmt_str(char *out,const char *in) {
+size_t fmt_str(char *out,const char *in) {
   register char* s=out;
   register const char* t=in;
   for (;;) {
