+++ fmt/fmt_ulong0.c
@@ -1,6 +1,6 @@
 #include "fmt.h"
 
-unsigned int fmt_ulong0(char *dest,unsigned long i,unsigned int pad) {
+size_t fmt_ulong0(char *dest,unsigned long i,size_t pad) {
   register unsigned int len;
   register unsigned long tmp;
   /* first count the number of bytes needed */
